Adds Light sensor based switching

+++ include/main.h
@@ -14,7 +14,7 @@ extern byte mode;
 extern uint8_t min;
 extern uint8_t hour;
 extern uint8_t sun;
-extern uint16_t light;
+extern uint8_t light;
 extern byte light_sensor;
 
 #endif
